requirement.txt
evdev==1.9.2
filelock==3.13.1
fsspec==2024.6.1
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
six==1.17.0
sympy==1.13.1
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
typing_extensions==4.9.0